Fix SaveAllStatusEffects From Removing Effects During Persist State

+++ src/map/status_effect_container.cpp
@@ -1632,7 +1632,7 @@ void CStatusEffectContainer::LoadStatusEffects()
  *                                                                       *
  ************************************************************************/
 
-void CStatusEffectContainer::SaveStatusEffects(bool logout)
+void CStatusEffectContainer::SaveStatusEffects(bool logout, bool skipRemove)
 {
     XI_DEBUG_BREAK_IF(m_POwner->objtype != TYPE_PC);
 
@@ -1640,7 +1640,7 @@ void CStatusEffectContainer::SaveStatusEffects(bool logout)
 
     for (CStatusEffect* PStatusEffect : m_StatusEffectSet)
     {
-        if ((logout && PStatusEffect->GetFlag() & EFFECTFLAG_LOGOUT) || (!logout && PStatusEffect->GetFlag() & EFFECTFLAG_ON_ZONE))
+        if (!skipRemove && ((logout && PStatusEffect->GetFlag() & EFFECTFLAG_LOGOUT) || (!logout && PStatusEffect->GetFlag() & EFFECTFLAG_ON_ZONE)))
         {
             RemoveStatusEffect(PStatusEffect, true);
             continue;
@@ -1700,7 +1700,11 @@ void CStatusEffectContainer::SaveStatusEffects(bool logout)
                        std::chrono::duration_cast<std::chrono::seconds>(PStatusEffect->GetStartTime().time_since_epoch()).count());
         }
     }
-    DeleteStatusEffects();
+
+    if (!skipRemove)
+    {
+        DeleteStatusEffects();
+    }
 }
 
 /************************************************************************
